+++ src/factory.h
@@ -11,7 +11,7 @@
 #include "taskqueue.h"
 
 typedef struct { // 火车模型
-    size_t _len; // 车头，存放车身的长度
+    int _len; // 车头，存放车身的长度
     char _buf[1024]; // 车身，存放要传输的数据
 } Train_t, * pTrain_t;
 
@@ -22,11 +22,11 @@ typedef struct {
     pthread_cond_t _cond;
     pthreadHandler_t _pThreadHandler;
     pthread_t * _pThreadId;
-    size_t _numThread;
+    int _numThread;
     short _flagStart;
 } Factory_t, * pFactory_t;
 
-void factoryInit(pFactory_t, size_t, pthreadHandler_t);
+void factoryInit(pFactory_t, int, pthreadHandler_t);
 void factoryStart(pFactory_t);
 
 #endif
